pyproject.toml
[project]
name = "mmgraphrag-odsc-west-2025"
version = "0.1.0"
description = "Multimodal GraphRAG (mmGraphRAG); Associative Search and Retrieval Only."
authors = [
    {name = "David Hughes", email = "david@hyperdimensionalcomputing.ai"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "jupyterlab>=4.3.5",
    "pillow>=11.1.0",
    "baml-py>=0.211.2",
    "torch>=2.6.0", # was "torch>=2.4.0,<2.5.0"
    "transformers>=4.48.2",
    "tqdm>=4.67.1",
    "tensorboard>=2.15.0,<2.16.0",
    "tensorflow>=2.15.1,<2.16.0",
    "ipywidgets>=8.1.5",
    "pydantic>=2.0.0",
    "soundfile>=0.12.1",
    "IPython>=9.6.0",
    "python-dotenv>=1.1.1",
    "pandas>=2.3.3",
    "lancedb>=0.25.2",
    "requests>=2.31.0",
    "numpy>=1.24.0,<2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mmgraphrag_odsc_west_2025"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
